renamed to singular
